Prettier and combined calls into account controller
